added stress

+++ 02__installconfig.sh
@@ -1,2 +1,7 @@
 #02__installconfig.sh
-rsync -av PPTARMG_config.sample/ ~/PPTARMG_config/
+if [ ! -e ~/PPTARMG_config ]; then
+	echo "PPTARMG_config not exists: copying sample config"
+	rsync -av PPTARMG_config.sample/ ~/PPTARMG_config/
+else
+	echo "PPTARMG_config already exists: not copying sample config"
+fi
